// constants/supabase-types.ts
export type Database = {
  al_watania: {
    Tables: {
// ...
      supervisors: {
        Row: {
          created_at: string
          details: string | null
          id: number
          name: string
          phone_numbers: number
          u_id: string
        }
        Insert: {
          created_at?: string
          details?: string | null
          id?: number
          name: string
          phone_numbers?: number
          u_id: string
        }
        Update: {
          created_at?: string
          details?: string | null
          id?: number
          name?: string
          phone_numbers?: number
          u_id?: string
        }
        Relationships: []
      }
    }
// ...
  }
// ...
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never